+++ tests/test_checker/test_noqa.py
@@ -62,6 +62,7 @@ def test_noqa_fixture_disabled(absolute_path, all_violations):
         'Z214': 0,
         'Z215': 1,
         'Z216': 0,
+        'Z217': 1,
 
         'Z220': 1,
         'Z221': 2,