+++ wemake_python_styleguide/options/defaults.py
@@ -65,3 +65,6 @@ MAX_DECORATORS: Final = 5
 
 #: Maximum number of same string usage in code.
 MAX_STRING_USAGES: Final = 3
+
+#: Maximum number of ``await`` expressions for functions or methods.
+MAX_AWAITS = 5  # the same as returns